temp: partial fix for update 0.220.3
Following update 0.220.3, `PrivilegeManager` and associated code has
been replaced by a new `PlatformManager` mechanism. Had a quick look
to fix stuff on my end and I figured I could come up with something for
`Guilds`, sending you the patch in case it may be useful.

Aside from that, `m_host` is not available anymore on `PlayerInfo`, but
I couldn't quite find a suitable replacement to identify a player from a
`ZRpc`, so this is not a complete fix... 😅